update: move helper function for events to a separate directory in the src directory
