+++ src/Events/events/linkedMessageRemove.ts
@@ -3,7 +3,7 @@ import BotEvent from "../BotEvent";
 import { channelIds } from "../../constants";
 import { Services } from "../../Services";
 import * as logger from "../../logger";
-import { isTHHorDevServer } from "../HelperFunction";
+import { isTHHorDevServer } from "../../Helper/EventsHelper";
 
 export default class LinkedMessageRemoveEvent extends BotEvent {
     public eventName = Events.MessageDelete;
